App Connectors: export createAppConnector() from connect-core
So that connect-core can do as much as possible to simplify the
implementation of an app connector.

createAppConnector() creates a function ready to be exported by the app
connector package, like connectVoting().

It passes to its callback an object providing contextual information
about the connection, the current app, the ipfs provider, verbosity, etc.

It waits for the app if a promise is provided, and emits an error if the
type is incorrect.

It handles the possibility to pass a connector name, or a connector name
and its configuration, as a secondary parameter.

It extends the passed object with the original App.

Lastly, it forwards the types of the configuration and the extended app,
if declared.