Add pollInterval to the options of TheGraph connectors
pollInterval is an option that can be set on all the TheGraph
connectors.

This commit also adds the possibility to consume the
organization connector from any app connector, allowing to retrieve its
options for example. This was done so that app connectors using TheGraph
can inherit from pollInterval defined in the main connector.