Replace GraphQL subscriptions (WS) by queries with polling (HTTP) (#203)
The Graph is having issues with GraphQL subscriptions, so we are temporarily
moving to queries with polling for the time being. Once subscriptions get
enabled again on The Graph, we will move back to GraphQL subscriptions.

This commit also adds the possibility to consume the organization connector
from any app connector, allowing to retrieve its options for example. This was
done so that app connectors using TheGraph can inherit from the pollInterval
defined in the main connector.

This commit also updates the nodejs/ demos.