Connected apps: use Proxy to inherit from both prototypes (#297)
This is needed to inherit from the prototypes of both App and
ConnectedApp. Another solution could be to declare all the App methods
on the object itself, but it would only make it possible to override
ConnectedApp with App methods, rather than the other way.

This is how it was done up to this point, but with this commit,
ConnectedApp now takes priority over App instead of the other way.
I think it makes most sense, since App can be accessed in another way,
and it will allow more flexibility when adding new properties to App.

Another thing to note is that using Proxy makes the object inspection
impossible. This is why two properties, _app and _connectedApp, have been
added on the proxy instance for debugging purposes.

Finally, a WeakMap is used internally to ensure that bind() is only called once
and doesn’t break comparisons by reference.