React API: useConnect() (#294)
## useConnect()

`useConnect()` is the new, unified way to use Connect through its React API. This new hook allows to fetch data using promises or subscriptions, which are the two ways Connect provides to consume data.

The main advantage is that it allows to interact with the standard Connect API, instead of having to rely on a set of hooks and `createAppHook()`. `useConnect()` can get called without any parameter, in which case it will return the organization object, the same way `connect()` does it for the Connect API.

// Before, with useOrganization()
function App() {
  const [org, { loading, error }] = useOrganization()
}

// With useConnect()
function App() {
  const [org, { loading, error }] = useConnect()
}

Up to this point, `useConnect()` behaves exactly as `useOrganization()`. But when it becomes interesting is that a callback can be passed as a parameter. This callback can return a promise, a subscription (`SubscriptionStart`), or any other data type. It also takes a single parameter: the currently loaded organization. `useConnect()` will take care of loading the data no matter what is being passed, and receive subscription updates. If an error gets thrown in any asynchronous operation, `error` will receive it.

// Load an app once, with org.app()
function App() {
  const [voting] = useConnect((org) => org.app('voting'))
}

// Subscribe to an app to receive updates, with org.onApp()
function App() {
  const [voting] = useConnect((org) => org.onApp('voting'))
}

Note: every subscription method exposed by Connect allows the callback to not be passed. In that case, a `SubscriptionStart` type will be returned. This is a function that can get used to start the subscription as a second step. The React API of Connect takes advantage of it by handling the subscriptions for you. This is why in the previous example, we use `org.onApp()` rather than `org.app()`. Remember to always use subscription methods if you want to receive updates with `useConnect()`.

`useConnect()` can also take an array of dependencies as a second parameter. It behaves exactly the same as the dependencies array passed to `useEffect()` or `useMemo()`, and will handle the process of restarting the subscription automatically.

function App() {
  const [appId, setAppId] = useState('voting')

  // The subscription will restart every time appId changes.
  const [app] = useConnect((org) => org.onApp(appId), [appId])

  return (
    <input
      value={appId}
      onChange={(event) => {
        setAppId(event.target.value)
      }}
    />
  )
}

Finally, app connectors can also get called directly in this hook. App connectors have been updated to support subscriptions.

import { useConnect } from '@aragon/connect-react'
import connectVoting from '@aragon/connect-voting'

function App() {
  const [voting] = useConnect((org) => connectVoting(org.onApp('voting')))
  const [votes] = useConnect(() => voting.onVotes(), [voting])

  return (
    <ul>
      {votes?.map((vote) => (
        <li>{vote.id}</li>
      ))}
    </ul>
  )
}

The reason why the `org` parameter is provided is for convenience. The dependencies array could be used to do the same thing, but using `org` is so common that passing it makes things easier.

// Using the dependencies array rather than the org parameter
function App() {
  const [org] = useConnect()
  const [voting] = useConnect(() => org.onApp('voting'), [org])
}

// Using the org parameter
function App() {
  const [voting] = useConnect((org) => org.onApp('voting'))
}

## Deprecated

Having `useConnect()` allows us to deprecate the other hooks and `createAppHook()`. This will make the React API of Connect not only much smaller, but also closer to the Connect API. Here is the list of deprecated functions and how to replace them.

### createAppHook()

// deprecated
const useVoting = createAppHook(connectVoting)
function App() {
  const [voting] = useApp('voting')
  const [votes] = useVoting(voting, (app) => app.onVotes())
}

// do this instead
function App() {
  const [voting] = useConnect((org) => connectVoting(org.onApp('voting')))
  const [votes] = useConnect(() => voting.onVotes(), [voting])
}

### useOrganization()

// deprecated
function App() {
  const [org] = useOrganization()
}

// do this instead
function App() {
  const [org] = useConnect()
}

### usePermissions()

// deprecated
function App() {
  const [permissions] = usePermissions()
}

// do this instead
function App() {
  const [permissions] = useConnect((org) => org.onPermissions())
}

### useApp()

// deprecated
function App() {
  const [app] = useApp('voting')
}

// do this instead
function App() {
  const [app] = useConnect((org) => org.onApp('voting'))
}

### useApps()

// deprecated
function App() {
  const [apps] = useApps('voting')
}

// do this instead
function App() {
  const [apps] = useConnect((org) => org.onApps('voting'))
}

## Changes

- Add a generic `useConnect()` hook, that transforms async calls and subscriptions into a result + status tuple.
- Split index.ts into separate files.
- `createAppHook()` is now a wrapper around `useConnectData()` that only loads the app.
- Improve memoization and loading states.
- Use `useConnect()` internally, remove `useSubscription()`.
- Make `Organization#onApp()`, `Organization#onApps()` and `Organization#onPermissions()` follow the new subscription pattern with optional callback.